moved lists to movies fragment
