+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <item android:id="@+id/home_nav"
-        android:title="Home"
+    <item android:id="@+id/featured_nav"
+        android:title="Featured"
         android:icon="@drawable/home_icon"/>
 
     <item android:id="@+id/movies_nav"
