+++ app/src/main/res/values/colors.xml
@@ -4,7 +4,7 @@
     <color name="colorPrimaryDark">#212121</color>
     <color name="colorAccent">#FFFFFF</color>
     <color name="colorAccentLight">#9E9E9E</color>
-    <color name="colorBackground">#212121</color>
+    <color name="colorBackground">#000000</color>
     <color name="colorBackgroundLight">#424242</color>
 
 </resources>
